Damage Bug
Finally fixed the damage thing by creating a health manager that can be modified by different enemy units

+++ Assets/Scripts/Settings/MoneyUi.cs
@@ -7,7 +7,7 @@ public class MoneyUi : MonoBehaviour
     public AudioSource audioSource;
     public AudioClip coinCollectSound;
 
-    private int lastMoneyAmount;
+    private float lastMoneyAmount;
 
     void Start()
     {
