+++ Assets/Scripts/Settings/PlayerStats.cs
@@ -6,8 +6,8 @@ using TMPro;
 
 public class PlayerStats : MonoBehaviour
 {
-    public static int Money;
-    public int StartMoney = 200;
+    public static float Money;
+    public float StartMoney = 200;
 
     public static int Lives = 5;
     public int startLives = 5;
@@ -146,7 +146,7 @@ public class PlayerStats : MonoBehaviour
     }
 
     //Debugging, remove later
-    public void AddMoneyNow(int amount)
+    public void AddMoneyNow(float amount)
     {
         Money += amount;
         Debug.Log($"Money added: {amount}. Total Money: {Money}");
